backend/requirements.txt
# ─── Django Core ───────────────────────────────────────────────────────────────
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0

# ─── Database ──────────────────────────────────────────────────────────────────
dj-database-url==2.1.0
psycopg2-binary==2.9.9

# ─── Environment ───────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ─── ML / Data Science ─────────────────────────────────────────────────────────
scikit-learn==1.3.2
xgboost==2.0.2
imbalanced-learn==0.11.0
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# ─── AI / Chatbot ──────────────────────────────────────────────────────────────
groq==0.4.2

# ─── File Handling ─────────────────────────────────────────────────────────────
Pillow==10.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
pypdf==4.2.0

# ─── Production Server ─────────────────────────────────────────────────────────
gunicorn==21.2.0
whitenoise==6.6.0

# ─── Utilities ─────────────────────────────────────────────────────────────────
requests==2.31.0
